pkg/slackUtils/requirements.txt
Flask
slack_sdk
python-dotenv